Add basic attack action and enable Characters to attack
The selectedCharacter can now attack enemies by clicking
on them, which uses their defaultAction. The attack so far
consists of queuing the attack action and walking up to
them and performing the empty action.

Note: when a unit passes through another unit (i.e.
occupies the same grid square) it overwrites the other
entity on the grid, erasing it. This will be mediated
by pathfinding to walk around enemies. I might need
to address it if we want friendly units to be able to
pass through each other.