Synchronize all threads
Due to the added complexity of dealing with synchronization, this
commit moves to a simpler model. Previously, access to every
individual variable or object was synchronized based on dependencies.
Now, for simpler development, threads will be synchronized such that
the game will either be handling input, rendering, or updating
game state at one time.